+++ tests/log_parser/test_crossreference_error_lines.py
@@ -19,8 +19,7 @@ def test_crossreference_error_lines(test_job):
     create_failure_lines(test_job, lines)
     create_text_log_errors(test_job, lines)
 
-    call_command('crossreference_error_lines', test_job.repository.name,
-                 test_job.guid)
+    call_command('crossreference_error_lines', str(test_job.id))
 
     summary = TextLogSummary.objects.all()
     assert len(summary) == 1
@@ -55,8 +54,7 @@ def test_crossreference_error_lines_truncated(test_job):
                          lines[:-1] + [({"action": "truncated"}, {})])
     create_text_log_errors(test_job, lines)
 
-    call_command('crossreference_error_lines', test_job.repository.name,
-                 test_job.guid)
+    call_command('crossreference_error_lines', str(test_job.id))
 
     summary_lines = TextLogSummaryLine.objects.all()
     assert len(summary_lines) == len(lines)
@@ -64,7 +62,6 @@ def test_crossreference_error_lines_truncated(test_job):
 
 
 def test_crossreference_error_lines_missing(test_job):
-
     lines = [(test_line, {}),
              (test_line, {"subtest": "subtest2"}),
              (test_line, {"status": "TIMEOUT"}),
@@ -75,8 +72,7 @@ def test_crossreference_error_lines_missing(test_job):
     create_failure_lines(test_job, lines[1:])
     create_text_log_errors(test_job, lines)
 
-    call_command('crossreference_error_lines', test_job.repository.name,
-                 test_job.guid)
+    call_command('crossreference_error_lines', str(test_job.id))
 
     failure_lines = FailureLine.objects.all()
     summary_lines = TextLogSummaryLine.objects.all()
