+++ treeherder/autoclassify/matchers.py
@@ -11,7 +11,7 @@ from django.conf import settings
 from django.db.models import Q
 from elasticsearch_dsl.query import Match as ESMatch
 
-from treeherder.autoclassify.management.commands.autoclassify import AUTOCLASSIFY_GOOD_ENOUGH_RATIO
+from treeherder.autoclassify.autoclassify import AUTOCLASSIFY_GOOD_ENOUGH_RATIO
 from treeherder.model.models import (ClassifiedFailure,
                                      FailureLine,
                                      FailureMatch,
